取得率の設定云々

+++ Assets/Script/Utility/SpRate.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public class SpRate : ScriptableObject
 {
-	public float[] spRates = new float[5];
+	public int[] spRates = new int[5];
 }
